requirements.txt
ultralytics
opencv-contrib-python
pyserial
numpy
torch
pyyaml
